Enable disabling network throttling.
Due to some oddness of how chrome handles h2 and h2 push when network
throttling is enabled, it is useful to be able to disable this.

To accomplish this goal, I added a new flag to the CLI, and then
modified the driver classes to take an options value and store it on the
class instance. The options are proxied down from the CLI flags ... into
the core lighthouse module and then directly into cri driver
constructor.

ref #646